refactor: DF-271 - Pull out event mappers
