+++ src/messaging/publish.js
@@ -1,14 +1,10 @@
-import {
-  AuditEventMessageCategory,
-  AuditEventMessageType,
-  messageSchema
-} from '@defra/forms-model'
+import { messageSchema } from '@defra/forms-model'
 import Joi from 'joi'
 
 import {
-  createFormMessageDataBase,
-  createV1MessageBase
-} from '~/src/messaging/mappers.js'
+  formCreatedEventMapper,
+  formTitleUpdatedMapper
+} from '~/src/messaging/mappers/form-events.js'
 import { publishEvent } from '~/src/messaging/publish-base.js'
 
 /**
@@ -29,59 +25,19 @@ async function publishFormEvent(auditMessage) {
  * @param {FormMetadata} metadata
  */
 export async function publishFormCreatedEvent(metadata) {
-  /** @type {FormCreatedMessageData} */
-  const data = {
-    formId: metadata.id,
-    slug: metadata.slug,
-    title: metadata.title,
-    organisation: metadata.organisation,
-    teamName: metadata.teamName,
-    teamEmail: metadata.teamEmail
-  }
+  const auditMessage = formCreatedEventMapper(metadata)
 
-  const auditMessageBase = createV1MessageBase(metadata)
-
-  return publishFormEvent({
-    ...auditMessageBase,
-    data,
-    category: AuditEventMessageCategory.FORM,
-    type: AuditEventMessageType.FORM_CREATED
-  })
+  return publishFormEvent(auditMessage)
 }
 
 /**
  * @param {FormMetadata} metadata
  * @param {FormMetadata} oldMetadata
- * @returns {Promise<PublishCommandOutput>}
  */
 export async function publishFormTitleUpdatedEvent(metadata, oldMetadata) {
-  const { title } = metadata
-  const { title: oldTitle } = oldMetadata
+  const auditMessage = formTitleUpdatedMapper(metadata, oldMetadata)
 
-  const baseData = createFormMessageDataBase(metadata)
-
-  /**
-   * @type {FormTitleUpdatedMessageData}
-   */
-  const data = {
-    ...baseData,
-    changes: {
-      previous: {
-        title: oldTitle
-      },
-      new: {
-        title
-      }
-    }
-  }
-  const auditMessageBase = createV1MessageBase(metadata)
-
-  return publishFormEvent({
-    ...auditMessageBase,
-    data,
-    category: AuditEventMessageCategory.FORM,
-    type: AuditEventMessageType.FORM_TITLE_UPDATED
-  })
+  return publishFormEvent(auditMessage)
 }
 
 /**
